src/pyproject.toml
[tool.poetry]
name = "my-flask-app"
version = "1.0.0"
description = "A Flask web application"
authors = ["Oscar Andres Macias Narvaez <omaciasnarvaez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
numpy = "^2.1.1"
pandas = "^2.2.3"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
pika = "^1.3.2"
psycopg2-binary = "^2.9.9"  # Cambiar a psycopg2-binary
safety = "^3.2.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"

[tool.poetry.group.dev.dependencies]
safety = "^3.2.7"
bandit = "^1.7.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
